Update footer design and change font settings
Updated the design and content of the footer in the src, dist and index.html, which now includes columns for a subscription section and a navigation area. The font has been changed from 'Poppins' to 'Sola' across several files and the font settings have been adjusted in tailwind.config.js and _tailwind.scss files.