Add pagination component and Arrow-red icon
Created a new Pagination and Arrow-red icon in both SRC and DIST directories. Updated the styles in _main.scss to support the pagination component and added the new icon Arrow-red in _icon.scss. Also, the dropdown function in the script.html file has been renamed to data. Lastly, included the Pagination component in both the index.html files.