Add Contact Us page
Added a new HTML file, 'ContactUs.html', to the 'src' directory and an equivalent 'ContactUs.html' file to the 'dist' directory. This new page contains a contact form, providing users a means to send their inquiries or messages to the website's administrators. The form includes fields for 'Full Name', 'Email', 'Message Subject', and 'Message'. In addition, the page provides contact details—specifically, the telephone number and email address—for the users to utilize.