Add Preview page and update styles
A new Preview page was added to the website. Existing styles in CSS files were updated to match the new requirements. The Footer component was also updated across several files. Addition of some extra classes to the tailwind.config.js file were made for new styling needs. A link to the new Preview page was added in readme.md.